Fix signedness comparison errors
Address the type comparison errors when --enable-debug configure option is set:

play_conf_validate.c: In function ‘tlog_play_conf_validate_journal’:
play_conf_validate.c:150:27: error: comparison of integer expressions of different signedness:
 ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Werror=sign-compare]
             for (i = 0; i < json_object_array_length(obj); i++) {

+++ lib/tlog/rec_session_conf.c
@@ -295,7 +295,7 @@ tlog_rec_session_conf_get_shell(struct tlog_errs **perrs,
     char **argv = NULL;
     size_t argi;
     char *arg = NULL;
-    int i;
+    size_t i;
 
     /* Read the login flag */
     login = json_object_object_get_ex(conf, "login", &obj) &&
@@ -363,7 +363,7 @@ tlog_rec_session_conf_get_shell(struct tlog_errs **perrs,
         argv[argi++] = arg;
         arg = NULL;
     }
-    for (i = 0; i < json_object_array_length(args); i++, argi++) {
+    for (i = 0; i < (size_t)json_object_array_length(args); i++, argi++) {
         obj = json_object_array_get_idx(args, i);
         arg = strdup(json_object_get_string(obj));
         if (arg == NULL) {
